Fix swiper error(#107)
