+++ client/js/include/swiper-main.js
@@ -2,7 +2,6 @@ const mainSwiper = new Swiper('.swiper-recommend', {
   slidesPerView: 4,
   spaceBetween: 18,
   slidesPerGroup: 4,
-  loop: true,
   loopFillGroupWithBlank: true,
   navigation: {
     nextEl: '.swiper-recommend-button-next',
@@ -13,7 +12,7 @@ const mainSwiper = new Swiper('.swiper-recommend', {
     onlyInViewport: false,
   },
   lazy: {
-    loadPrevNext: true // 이전, 다음 이미지는 미리 로딩
+    loadPrevNext: true
   },
 
   a11y: {
@@ -27,7 +26,6 @@ const mainSwiper2 = new Swiper('.swiper-best-price', {
   slidesPerView: 4,
   spaceBetween: 18,
   slidesPerGroup: 4,
-  loop: true,
   loopFillGroupWithBlank: true,
   navigation: {
     nextEl: '.swiper-best-price-button-next',
@@ -36,7 +34,9 @@ const mainSwiper2 = new Swiper('.swiper-best-price', {
   lazy: {
     loadPrevNext: true
   },
-
+  keyboard: {
+    enabled: true,
+  },
   a11y: {
     containerRoleDescriptionMessage: '가격 할인 상품',
     nextSlideMessage: '다음 슬라이드 버튼 입니다',
@@ -44,32 +44,3 @@ const mainSwiper2 = new Swiper('.swiper-best-price', {
   }
 
 });
-
-
-const asideSwiper = new Swiper('.aside-swiper', {
-  direction: 'vertical',
-  slidesPerView: 3,
-  grid: {
-    column: 2,
-  },
-
-  spaceBetween: 10,
-  slidesPerGroup: 3,
-  loop: true,
-  loopFillGroupWithBlank: true,
-  navigation: {
-    nextEl: '.aside-swiper-button-next',
-    prevEl: '.aside-swiper-button-prev',
-  },
-  lazy: {
-    loadPrevNext: true
-  },
-
-  a11y: {
-    containerRoleDescriptionMessage: '최근 본 상품',
-    nextSlideMessage: '다음 슬라이드 버튼 입니다',
-    prevSlideMessage: '이전 슬라이드 버튼 입니다',
-  }
-});
-
-
