+++ client/page/home.html
@@ -29,6 +29,7 @@
   <script src="../js/include/cart_modal.js" type="module"></script>
   <script src="../js/include/swiper-main.js" type="module"></script>
   <script src="../js/include/swiper_main_banner.js" type="module"></script>
+  <script src="../js/include/swiper_aside.js" type="module"></script>
 </head>
 
 <body>
